patientList-BE/src/config.py
import os
from dotenv import load_dotenv


class Config:

    load_dotenv()

    SECRET_KEY = os.getenv('SECRET_KEY')
    # SQLALCHEMY_DATABASE_URI = os.getenv('SQLALCHEMY_DATABASE_URI')
    SQLALCHEMY_DATABASE_URI = os.getenv('SQLALCHEMY_DATABASE_URL')
    MAIL_SERVER = 'smtp.gmail.com'
    MAIL_PORT = 465
    MAIL_USE_TLS = False
    MAIL_USE_SSL = True
    MAIL_USERNAME = os.getenv('EMAIL_USER')
    MAIL_PASSWORD = os.getenv('EMAIL_PASS')
    FLASK_ENV = os.getenv('FLASK_ENV')
    CSRF_ENABLED = True

    print("GMAIL CREDS\n")
    print(MAIL_USERNAME)
    print(MAIL_PASSWORD)